Add base-calc.html
This example originally showed up in Introduction to JavaScript Programming with XML and PHP 1st edition (ISBN 9780133068306). The original example was a simple demonstration of the toString() method. I went further and decided to add more discretion to the example by giving the user the opportunity to specify a base of their own.

If a non-number is added, the program will simply say NaN. I haven't given doubles a try yet. But this is really meant for integers. I used this to double check my written conversion technique from my MAT 145 class, where we spent an entire chapter converting between base 2 and 10.